Move the unit tests out of the root package.
The karma harness declares `"@bedrock/vue-wallet": "file:.."` and runs under
`--preserve-symlinks`, so it resolves this package's imports through the ROOT
node_modules. Anything installed there is therefore part of the karma build's
module graph.

Adding vitest to the root devDependencies put `lru-cache` there:

  @vue/test-utils -> js-beautify -> glob -> path-scurry -> lru-cache@10.4.3
  jsdom -> lru-cache@11.5.2

`test/package.json` pins `@bedrock/web-profile-manager@^20.0.0`, which is a
v6-era package: it declares `lru-cache ^6.0.0` and imports the default export,
which is correct for v6 and absent from v10 and v11. With a v10/v11 copy
sitting in the root, that is what the test build resolved, and webpack failed
on two missing-export errors in node_modules. Nothing was wrong with any of
the packages involved; each pairs its import style with its declared range
correctly. Only the shadowing was wrong.

So the unit tests get their own package, the way the karma tests already do.
Nothing they need reaches the root, and the karma build resolves the same
tree main does. Specs move alongside their config rather than sitting beside
the source, matching `test/`.

This also stops the package publishing its own test scaffolding: `mocks/` had
accumulated ten vitest mocks next to the one file that is genuinely part of
the published surface, and the eight spec files shipped inside `components/`
and `lib/`. The tarball goes from 88 files to 70.

Run them with `cd test-unit && npm install && npm test`. They are not wired
into CI, which runs lint, karma and coverage only; wiring them in touches the
shared workflow and is left as a separate decision.

75 tests across 8 files pass unchanged; eslint clean.
