+++ package.json
@@ -46,18 +46,7 @@
   },
   "devDependencies": {
     "@digitalbazaar/eslint-config": "^7.0.0",
-    "@vitejs/plugin-vue": "^6.0.8",
-    "@vitest/ui": "^4.1.10",
-    "@vue/test-utils": "^2.4.11",
-    "@vueuse/core": "^10.9.0",
-    "eslint": "^9.38.0",
-    "happy-dom": "^20.11.1",
-    "jsdom": "^30.0.1",
-    "quasar": "^2.24.0",
-    "sass-embedded": "^1.100.0",
-    "vitest": "^4.1.10",
-    "vue": "^3.5.41",
-    "vue-router": "^5.2.0"
+    "eslint": "^9.38.0"
   },
   "engines": {
     "node": ">=18"
