+++ test-unit/lib/renderMethod.spec.js
@@ -2,7 +2,7 @@
  * Copyright (c) 2026 Digital Bazaar, Inc. All rights reserved.
  */
 import {afterEach, beforeEach, describe, expect, it, vi} from 'vitest';
-import {getRenderedImages} from '../renderMethod.js';
+import {getRenderedImages} from '../../lib/renderMethod.js';
 
 const DATA_URI = 'data:image/svg+xml;base64,PHN2Zy8+';
