+++ test-unit/mocks/bedrock/web-wallet.js
