+++ test-unit/mocks/digitalbazaar/vue-extendable-event.js
